Upgrade SQLAlchemy version from 1.4.46 to 2.0.7
1. Upgrade SQLAlchemy
2. Use mypy and solve related typing problems

+++ bot_ui/callbacks.py
@@ -1,3 +1,5 @@
+from typing import no_type_check
+
 from aiogram import Dispatcher, F, Bot
 from aiogram.exceptions import TelegramBadRequest
 from aiogram.types import CallbackQuery, Message
@@ -57,42 +59,51 @@ async def show_main_keyboard(key: StorageKey,
 
 
 async def show_tag_filter_keyboard(query: CallbackQuery, context: BotContext):
+    if query.message is None:
+        return
+
     key = StorageKey.from_callback_query(query)
     if data := await context.storage.get_data(key):
         data.back_callback_data = NavigationData(keyboard=Keyboard.MAIN).pack()
         data.tags_offset = 0
-        async with context.SessionMaker.begin() as session:
+        async with context.session_maker.begin() as session:
             keyboard = await build_tag_filter_keyboard(data.tags_offset,
                                                        MAX_TAG_COUNT,
                                                        data.tags_ids,
                                                        data.back_callback_data,
                                                        session)
-        await query.message.edit_text("Select tags for filter youtube channels:")
-        await query.message.edit_reply_markup(reply_markup=keyboard)
-        await context.storage.set_data(key, data)
+            await query.message.edit_text("Select tags for filter youtube channels:")
+            await query.message.edit_reply_markup(reply_markup=keyboard)
+            await context.storage.set_data(key, data)
 
 
 async def show_tg_keyboard(query: CallbackQuery, context: BotContext):
+    if query.message is None:
+        return
+
     key = StorageKey.from_callback_query(query)
     if data := await context.storage.get_data(key):
         data.back_callback_data = NavigationData(keyboard=Keyboard.MAIN).pack()
         data.tgs_offset = 0
-        async with context.SessionMaker.begin() as session:
+        async with context.session_maker.begin() as session:
             keyboard = await build_telegram_tg_keyboard(data.tgs_offset,
                                                         MAX_TG_COUNT,
                                                         data.back_callback_data,
                                                         session)
-        await query.message.edit_text('Telegram chats and threads:')
-        await query.message.edit_reply_markup(reply_markup=keyboard)
-        await context.storage.set_data(key, data)
+            await query.message.edit_text('Telegram chats and threads:')
+            await query.message.edit_reply_markup(reply_markup=keyboard)
+            await context.storage.set_data(key, data)
 
 
 async def show_channels_keyboard(query: CallbackQuery, context: BotContext):
+    if query.message is None:
+        return
+
     key = StorageKey.from_callback_query(query)
     if data := await context.storage.get_data(key):
         data.back_callback_data = NavigationData(keyboard=Keyboard.TAG_FILTER).pack()
         data.yt_channels_offset = 0
-        async with context.SessionMaker.begin() as session:
+        async with context.session_maker.begin() as session:
             assert data.original_chat_id  # FIXME
             keyboard = await build_channel_keyboard(data.original_chat_id,
                                                     data.original_thread_id,
@@ -102,25 +113,29 @@ async def show_channels_keyboard(query: CallbackQuery, context: BotContext):
                                                     data.tags_ids,
                                                     data.back_callback_data,
                                                     session)
-        await query.message.edit_text("YouTube channels:")
-        await query.message.edit_reply_markup(reply_markup=keyboard)
-        await context.storage.set_data(key, data)
+            await query.message.edit_text("YouTube channels:")
+            await query.message.edit_reply_markup(reply_markup=keyboard)
+            await context.storage.set_data(key, data)
 
 
 async def yt_channels_in_tg_pressed(query: CallbackQuery,
                                     callback_data: TgData,
                                     context: BotContext):
+    if query.message is None:
+        return
+
     key = StorageKey.from_callback_query(query)
     if data := await context.storage.get_data(key):
         data.original_chat_id = callback_data.chat_id
         data.original_thread_id = callback_data.thread_id
         data.back_callback_data = NavigationData(keyboard=Keyboard.TG_OBJECTS).pack()
-        async with context.SessionMaker.begin() as session:
+        async with context.session_maker.begin() as session:
             keyboard = await build_tag_filter_keyboard(data.tags_offset,
                                                        MAX_TAG_COUNT,
                                                        data.tags_ids,
                                                        data.back_callback_data,
                                                        session)
+            assert query.message
             await query.message.edit_text("YouTube channels:")  # TODO: For telegram
             await query.message.edit_reply_markup(reply_markup=keyboard)
         await context.storage.set_data(key, data)
@@ -129,11 +144,16 @@ async def yt_channels_in_tg_pressed(query: CallbackQuery,
 async def channel_checked(query: CallbackQuery,
                           callback_data: ChannelData,
                           context: BotContext):
+    if query.message is None:
+        return
+
     key = StorageKey.from_callback_query(query)
     if data := await context.storage.get_data(key):
-        async with context.SessionMaker.begin() as session:
+        async with context.session_maker.begin() as session:
             assert data.original_chat_id  # FIXME
-            if tg := await get_destinations(data.original_chat_id, data.original_thread_id, session):
+            if tg := await get_destinations(data.original_chat_id,
+                                            data.original_thread_id,
+                                            session):
                 if not callback_data.enabled:
                     await add_forwarding(callback_data.id,
                                          tg.chat.original_id,
@@ -153,6 +173,7 @@ async def channel_checked(query: CallbackQuery,
                                                         data.tags_ids,
                                                         data.back_callback_data,
                                                         session)
+                assert query.message
                 await query.message.edit_reply_markup(reply_markup=keyboard)
             await context.storage.set_data(key, data)
 
@@ -160,9 +181,11 @@ async def channel_checked(query: CallbackQuery,
 async def tag_button_pressed(query: CallbackQuery,
                              callback_data: TagFilterData,
                              context: BotContext):
+    if query.message is None:
+        return
     key = StorageKey.from_callback_query(query)
     if data := await context.storage.get_data(key):
-        async with context.SessionMaker.begin() as session:
+        async with context.session_maker.begin() as session:
             data.tags_ids ^= {callback_data.id, }
             keyboard = await build_tag_filter_keyboard(data.tags_offset,
                                                        MAX_TAG_COUNT,
@@ -170,46 +193,52 @@ async def tag_button_pressed(query: CallbackQuery,
                                                        data.back_callback_data,
                                                        session)
             await query.message.edit_reply_markup(reply_markup=keyboard)
-        await context.storage.set_data(key, data)
+            await context.storage.set_data(key, data)
 
 
 async def close_keyboard(query: CallbackQuery):
-    await query.message.delete()
+    if query.message:
+        await query.message.delete()
 
 
 async def back_to_main_keyboard(query: CallbackQuery, bot: Bot, context: BotContext):
-    await show_main_keyboard(StorageKey.from_callback_query(query),
-                             query.message,
-                             bot,
-                             context)
+    if query.message:
+        key = StorageKey.from_callback_query(query)
+        await show_main_keyboard(key, query.message, bot, context)
 
 
 async def attach_tags_callback(query: CallbackQuery,
                                callback_data: AttachTagData,
                                context: BotContext):
+    if not query.message:
+        return
+
     key = StorageKey.from_callback_query(query)
     if data := await context.storage.get_data(key):
         data.back_callback_data = NavigationData(keyboard=Keyboard.YT_CHANNELS).pack()
         data.tags_offset = 0
-        data.original_channel_id = callback_data.channel_id
-        async with context.SessionMaker.begin() as session:
+        data.channel_id = callback_data.channel_id
+        async with context.session_maker.begin() as session:
             if channel := await get_yt_channel_by_id(callback_data.channel_id, session):
-                keyboard = await build_attach_tags_keyboard(data.original_channel_id,
+                keyboard = await build_attach_tags_keyboard(data.channel_id,
                                                             data.tags_offset,
                                                             MAX_TAG_COUNT,
                                                             data.back_callback_data,
                                                             session)
                 await query.message.edit_text(f'Select tag for "{channel.title}"')
                 await query.message.edit_reply_markup(reply_markup=keyboard)
-        await context.storage.set_data(key, data)
+                await context.storage.set_data(key, data)
 
 
 async def yt_channel_tag_button_pressed(query: CallbackQuery,
                                         callback_data: YtChannelTagData,
                                         context: BotContext):
+    if not query.message:
+        return
+
     key = StorageKey.from_callback_query(query)
     if data := await context.storage.get_data(key):
-        async with context.SessionMaker.begin() as session:
+        async with context.session_maker.begin() as session:
             if callback_data.enabled:
                 await delete_yt_channel_tag(callback_data.tag_id, callback_data.channel_id, session)
             else:
@@ -226,9 +255,12 @@ async def yt_channel_tag_button_pressed(query: CallbackQuery,
 async def status_button_pressed(query: CallbackQuery,
                                 callback_data: StatusData,
                                 context: BotContext):
+    if not query.message:
+        return
+
     key = StorageKey.from_callback_query(query)
     if data := await context.storage.get_data(key):
-        async with context.SessionMaker.begin() as session:
+        async with context.session_maker.begin() as session:
             await set_telegram_chat_status(callback_data.chat_id,
                                            callback_data.status,
                                            session)
@@ -244,7 +276,7 @@ async def nav_button_pressed(query: CallbackQuery,
                              context: BotContext):
     key = StorageKey.from_callback_query(query)
     if data := await context.storage.get_data(key):
-        async with context.SessionMaker.begin() as session:
+        async with context.session_maker.begin() as session:
             match callback_data.keyboard:
                 case Keyboard.TAG_FILTER:
                     data.tags_offset = callback_data.offset
@@ -272,15 +304,17 @@ async def nav_button_pressed(query: CallbackQuery,
                                                             session)
                 case Keyboard.ATTACH_TAGS:
                     data.tags_offset = callback_data.offset
-                    keyboard = await build_attach_tags_keyboard(data.original_channel_id,
+                    keyboard = await build_attach_tags_keyboard(data.channel_id,
                                                                 data.tags_offset,
                                                                 MAX_TAG_COUNT,
                                                                 data.back_callback_data,
                                                                 session)
+            assert query.message
             await query.message.edit_reply_markup(reply_markup=keyboard)
         await context.storage.set_data(key, data)
 
 
+@no_type_check
 def register_callback_queries(dp: Dispatcher,
                               chat_admin_filter: ChatAdminFilter,
                               bot_admin_filter: BotAdminFilter):
